+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - added Dockerfile for production
 - [#19 - Table tags and inline styles can be used in ckEditor for email templates](https://github.com/shopsys/demoshop/pull/19)
     - `ivory_ck_editor.yml` is updated so email templates wysiwyg can parse table tags and inline styles for all elements
+    - datafixtures for email templates are added so demoshop has now customized emails
 
 ### Fixed
 - [#8 - Category now has second description attribute that is displayed on the product list page above the product list](https://github.com/shopsys/demoshop/pull/8)